refactor: extract booking screen logic into custom hook
- Move all booking screen logic from booking-screen.tsx to a new useBookingScreen hook in useBooking.tsx
- Create Stylist type and related interfaces
- Update Service interface to use string IDs and make duration optional
- Move mock stylist and service data from component to servicesData.tsx
- Streamline booking-screen.tsx component by consuming the custom hook
- Add TODO comment about potential further refactoring of booking logic

This refactor improves code organization, makes business logic more testable, and follows the custom hook pattern established in the codebase.
