+++ app/mock-data/servicesData.tsx
@@ -23,3 +23,18 @@ export const serviceMockData: Service[] = [
     description: 'Expert styling for special occasions or everyday looks',
   },
 ];
+
+// Mock data for services and stylists
+export const servicesBookingMockData = [
+  { id: '1', name: 'Haircut', price: 35 },
+  { id: '2', name: 'Coloring', price: 80 },
+  { id: '3', name: 'Styling', price: 50 },
+];
+
+export const stylistsMockData = [
+  { id: '1', name: 'John' },
+  { id: '2', name: 'Sarah' },
+  { id: '3', name: 'Michael' },
+  { id: '4', name: 'Jessica' },
+  { id: 'any', name: 'Any Available Stylist' },
+];
