+++ app/types/stylistTypes.tsx
@@ -1,0 +1,8 @@
+// Stylist type definition
+export interface Stylist {
+  id: string;
+  name: string;
+  specialties?: string[];
+  experience?: number;
+  bio?: string;
+}
